backend/requirements.txt
# Pitch-Sync Engine - Python Dependencies
# Versions pinned for reproducibility (PATTERN-001)

fastapi==0.115.0
uvicorn==0.32.0
boto3==1.35.0
pydantic==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
openai==1.50.0
sqlmodel==0.0.22
requests==2.32.0
Pillow==11.0.0

# Testing (TEST-001)
pytest==8.2.0
pytest-asyncio==0.24.0
httpx==0.27.0
